backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
marshmallow==3.20.1
python-dateutil==2.8.2
Werkzeug==2.3.7
python-dotenv==1.0.0
jsonschema==4.19.2
requests==2.31.0
PyJWT==2.8.0
PyPDF2==3.0.1
